server/requirements.txt
appdirs==1.4.3
asgi-redis==1.4.0
asgiref==1.1.2
attrs==17.2.0
autobahn==17.5.1
Automat==0.6.0
channels==1.1.3
click==6.7
constantly==15.1.0
daphne==1.2.0
deap==1.0.2
Django==1.11.1
incremental==17.5.0
JPype1==0.6.2
msgpack-python==0.4.8
numpy==1.12.1
packaging==16.8
pkg-resources==0.0.0
psycopg2==2.7.1
pur==4.0.2
pyparsing==2.2.0
pytz==2017.2
redis==2.10.5
six==1.10.0
SQLAlchemy==1.1.10
Twisted==17.1.0
txaio==2.7.1
zope.interface==4.4.1
